Minor improve vocabulary used in README
In fact the (x,y,ORIENTATION) vector is not the _location_ of
the robot, that's to say, it's not *only* the location of the robot,
but also it orientation. Now we'll be working with locations, it's
better to disambiguate both concepts.

From now on, (x,y) should be called location, (ORIENTATION) orientation,
and (x,y,ORIENTATION) should be called situation. IMO. I'll try to
progressively fix method and variable names in that way.