Fix method shadowing issue, complete move of output formatting responsibility
Damned! How could that go unnoticed?

Specs were obviously not clear enough to make their blind spots visible.
That may be fixed now (see below), but taking a backward look at it would be
a good thing to do.

The fist part of the responsibility move permitted to reduce ambiguity
in specs scope: Simulation specs are integration specs because that's
the model users will most directly interact with. Robot and Table specs
are more unit, detail-oriented specs. Even if that was my intention
from the start, it was quite unclear until the simulation specs rewrite
on occasion of that responsibility move.